requirements.txt
aiohttp==3.8.4 ; python_version >= "3.10" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4"
charset-normalizer==3.1.0 ; python_version >= "3.10" and python_version < "4"
discord-py==2.2.2 ; python_version >= "3.10" and python_version < "4.0"
frozenlist==1.3.3 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
multidict==6.0.4 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.10" and python_version < "4"
urllib3==1.26.15 ; python_version >= "3.10" and python_version < "4"
yarl==1.9.1 ; python_version >= "3.10" and python_version < "4.0"
